feat(repeating questions): use a skip back to make repeating questions
Whenever the code now sees a skip back, it duplicates all the previous questions, and asks them to the user again. This way, the user can answer all the questions as many times as they want. This will be useful for questions like, Job History, or Family History, when you want to ask some questions several times.