feat(back button): add back button
The front end now sends a reponse to the backend with an extra property 'viewed', which stores if the question has been viewed or not. When someone views a page, all those questions are marked as viewed. When a user clicks the back button, we find, the first viewed questions before this one - there might not unviewed questions if someone had skipped over them.