feat(history): add full support for history
Huge change. The main benefits are the back and next button working with all skip logic.

When a response is created, its made with all the responses in there. They are then marked as viewed or not viewed. These 'responseElements' are then used to build the questionnaire the user sees. The real elements are stored in the version, elements are like, questions, section headings, text information etc.