change weight to pounds
now its similar to feet and inches
also change pounds, number, feet, inches to be stored as integers, instead of strings. This will be useful for the logic later on.

Dates, no idea how to store it for comparisons. You can't parse it as a date, because, if someone is born on the 12 of Jan in Australia, their birth date doesn't become 11 of Jan in the US. This is what happens when you use unix time stamps to store dobs. Unless, you include the timezone. So, leaving this out for now.