add end page
The add page appears now at the top of the summary page.
The 'submit' button is now back again at the end of the questionnaire
not on the summary page. This was changed because it just makes it
easier to find where to put the end page. Since we changed the
questionnaire to allow you to scroll page, a user can review all their
answers before submitting that way. We dont need a separate page
anymore. Thus the summary page, is a post submission summary page now.

https://github.com/genome-one/sanford/issues/194