fix shared cookie between builder and form
The builder and form are running on the same domain on the server
so they are sharing cookies. This is causing problems. So I've renamed
the cookie so they dont clash, and, if we have enough info to auth
which is userId, consentTypeId and timestamp, then I throw out
any old JWT, just so there are no problems.

Close #25