fix logic causing question not to become invisible
the problem:
selecting an answer that made a question appear, and then changing
that answer to one that made the question disappear, didnt make the
question disappear.

cause: we were looking for answerId in our store, but one the answer
had been selected the first time, it was never removed from the store,
so it was always present.

fix: we needed to check all the responseElements, to see if our answer
was attached to any of them.

Closes #84