🆕Enable to enlarge font in form (#99)
Problem: Sanford wants to cater for users with poor eyesight and enable
them to enlarge the font on the questionnaire form. They also want this
to match up with the styling in SecureConsent.

Solution: Added a button that toggles through 3 different font sizes - original, larger, largest and then back to original. This increases the font size in text and inside form elements (like text areas).

Closes: https://github.com/GenomeOne/web-questionnaire-form/issues/93