change none of the above behaviour
Problem: We had a lot of checking for checking if ‘none of the above’ was actually the last answer. In reality, for what Sanford whats, ‘None of the above’ basically equates to ‘None of the answers’.

Solution: So we can simplify the code a bit, so that whenever someone clicks ‘none of the above’ no matter if its the first answer, or the last one, we just clear all the answers. It’s not exactly what “None of the above” means in a literal sense, but it’s how its being used.

Close #136